project/starter/requirements.txt
google-adk>=1.13.0
google-cloud-discoveryengine>=0.13.11
toolbox-core>=0.5.0
